0.2.0 Removed dependency on clmnis
Removed the dependency on clmnis for information on members. Now only very basic information is returned based on content available in the Votes API. clmnis can still be used for more detailed information on members by joining on mnis_id. Added a new variable for lords divisions which aggregates votes on a division by lord type.